chore: move contract files to contracts folder
