+++ packages/contract/src/__test__/call-test-contract/call-test-contract.test.ts
@@ -6,8 +6,8 @@ import { TestUtils } from '@fuel-ts/wallet';
 import { readFileSync } from 'fs';
 import { join } from 'path';
 
-import type Contract from '../../contract';
-import ContractFactory from '../../contract-factory';
+import type Contract from '../../contracts/contract';
+import ContractFactory from '../../contracts/contract-factory';
 
 import abiJSON from './out/debug/call-test-abi.json';
