+++ packages/contract/src/index.ts
@@ -1,4 +1,3 @@
-export * from './contract';
-export { default as ContractFactory } from './contract-factory';
-export { default as Contract } from './contract';
+export { default as ContractFactory } from './contracts/contract-factory';
+export { default as Contract } from './contracts/contract';
 export * as ContractUtils from './util';
